Potential panic when using manual ACK
Measures were added in 2021 to ensure that matchAndDispatch waited for handlers to complete
before exiting. Unfortunately this did not fully account for manual acknowledgments (which could
be triggered at any time). This change simplifies Ack handling to avoid the issue (includes a test). Note that this is not a complete solution (ignoring an ACK after the connection has dropped will not always be the right option).

Ref issue #726